requirements.txt
google-api-python-client>=2.125.0,<3
google-auth-oauthlib>=1.2.0,<2
google-auth-httplib2>=0.2.0,<1
PyYAML>=6.0.1,<7
openai>=1.40.0,<2
yt-dlp>=2024.8.6
python-telegram-bot>=20.0,<21
openai-whisper>=20231117
fastapi>=0.104.0,<1
uvicorn[standard]>=0.24.0,<1
jinja2>=3.1.0,<4
python-multipart>=0.0.6,<1
pytz>=2023.3
google-cloud-vision>=3.4.0
httpx>=0.24.1,<0.25
aiohttp>=3.8.0

# Testing
pytest>=7.4.0,<9
pytest-cov>=4.1.0,<6
respx==0.20.*
